+++ android/app/src/main/java/com/app/edonymyeon/data/datasource/consumptions/ConsumptionsRemoteDataSource.kt
@@ -1,8 +1,8 @@
 package com.app.edonymyeon.data.datasource.consumptions
 
-import com.app.edonymyeon.data.common.ApiResponse
 import com.app.edonymyeon.data.dto.response.ConsumptionsResponse
 import com.app.edonymyeon.data.service.ConsumptionsService
+import com.app.edonymyeon.data.service.client.calladapter.ApiResponse
 import javax.inject.Inject
 
 class ConsumptionsRemoteDataSource @Inject constructor(
