+++ android/app/src/main/java/com/app/edonymyeon/data/datasource/post/PostDataSource.kt
@@ -1,11 +1,11 @@
 package com.app.edonymyeon.data.datasource.post
 
-import com.app.edonymyeon.data.common.ApiResponse
 import com.app.edonymyeon.data.dto.request.PostEditorRequest
 import com.app.edonymyeon.data.dto.response.CommentsResponse
 import com.app.edonymyeon.data.dto.response.PostDetailResponse
 import com.app.edonymyeon.data.dto.response.PostEditorResponse
 import com.app.edonymyeon.data.dto.response.Posts
+import com.app.edonymyeon.data.service.client.calladapter.ApiResponse
 import java.io.File
 
 interface PostDataSource {
