+++ android/app/src/main/java/com/app/edonymyeon/data/datasource/preference/PreferenceDataSource.kt
@@ -1,8 +1,8 @@
 package com.app.edonymyeon.data.datasource.preference
 
-import com.app.edonymyeon.data.common.ApiResponse
 import com.app.edonymyeon.data.dto.request.NotificationPreferenceRequest
 import com.app.edonymyeon.data.dto.response.NotificationPreferenceResponse
+import com.app.edonymyeon.data.service.client.calladapter.ApiResponse
 
 interface PreferenceDataSource {
     suspend fun getNotificationPreference(): ApiResponse<NotificationPreferenceResponse>
