+++ android/app/src/main/java/com/app/edonymyeon/data/datasource/preference/PreferenceRemoteDataSource.kt
@@ -1,9 +1,9 @@
 package com.app.edonymyeon.data.datasource.preference
 
-import com.app.edonymyeon.data.common.ApiResponse
 import com.app.edonymyeon.data.dto.request.NotificationPreferenceRequest
 import com.app.edonymyeon.data.dto.response.NotificationPreferenceResponse
 import com.app.edonymyeon.data.service.PreferenceService
+import com.app.edonymyeon.data.service.client.calladapter.ApiResponse
 import javax.inject.Inject
 
 class PreferenceRemoteDataSource @Inject constructor(
