+++ android/app/src/main/java/com/app/edonymyeon/data/datasource/profile/ProfileDataSource.kt
@@ -1,12 +1,12 @@
 package com.app.edonymyeon.data.datasource.profile
 
-import com.app.edonymyeon.data.common.ApiResponse
 import com.app.edonymyeon.data.dto.WriterDataModel
 import com.app.edonymyeon.data.dto.request.ProfileUpdateRequest
 import com.app.edonymyeon.data.dto.request.PurchaseConfirmRequest
 import com.app.edonymyeon.data.dto.request.SavingConfirmRequest
 import com.app.edonymyeon.data.dto.response.AuthDuplicateResponse
 import com.app.edonymyeon.data.dto.response.MyPostsResponse
+import com.app.edonymyeon.data.service.client.calladapter.ApiResponse
 import java.io.File
 
 interface ProfileDataSource {
