+++ android/app/src/main/java/com/app/edonymyeon/data/datasource/profile/ProfileRemoteDataSource.kt
@@ -1,6 +1,5 @@
 package com.app.edonymyeon.data.datasource.profile
 
-import com.app.edonymyeon.data.common.ApiResponse
 import com.app.edonymyeon.data.dto.WriterDataModel
 import com.app.edonymyeon.data.dto.request.ProfileUpdateRequest
 import com.app.edonymyeon.data.dto.request.PurchaseConfirmRequest
@@ -8,6 +7,7 @@ import com.app.edonymyeon.data.dto.request.SavingConfirmRequest
 import com.app.edonymyeon.data.dto.response.AuthDuplicateResponse
 import com.app.edonymyeon.data.dto.response.MyPostsResponse
 import com.app.edonymyeon.data.service.ProfileService
+import com.app.edonymyeon.data.service.client.calladapter.ApiResponse
 import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.MultipartBody
 import okhttp3.RequestBody.Companion.asRequestBody
