+++ android/app/src/main/java/com/app/edonymyeon/data/datasource/report/ReportDataSource.kt
@@ -1,7 +1,7 @@
 package com.app.edonymyeon.data.datasource.report
 
-import com.app.edonymyeon.data.common.ApiResponse
 import com.app.edonymyeon.data.dto.request.ReportRequest
+import com.app.edonymyeon.data.service.client.calladapter.ApiResponse
 
 interface ReportDataSource {
     suspend fun postReport(
