+++ android/app/src/main/java/com/app/edonymyeon/data/datasource/report/ReportRemoteDataSource.kt
@@ -1,8 +1,8 @@
 package com.app.edonymyeon.data.datasource.report
 
-import com.app.edonymyeon.data.common.ApiResponse
 import com.app.edonymyeon.data.dto.request.ReportRequest
 import com.app.edonymyeon.data.service.ReportService
+import com.app.edonymyeon.data.service.client.calladapter.ApiResponse
 import javax.inject.Inject
 
 class ReportRemoteDataSource @Inject constructor(
