+++ android/app/src/main/java/com/app/edonymyeon/data/datasource/search/SearchDataSource.kt
@@ -1,7 +1,7 @@
 package com.app.edonymyeon.data.datasource.search
 
-import com.app.edonymyeon.data.common.ApiResponse
 import com.app.edonymyeon.data.dto.response.Posts
+import com.app.edonymyeon.data.service.client.calladapter.ApiResponse
 
 interface SearchDataSource {
     suspend fun getSearchResult(query: String, page: Int): ApiResponse<Posts>
