+++ android/app/src/main/java/com/app/edonymyeon/data/service/NotificationService.kt
@@ -1,7 +1,7 @@
 package com.app.edonymyeon.data.service
 
-import com.app.edonymyeon.data.common.ApiResponse
 import com.app.edonymyeon.data.dto.response.NotificationsResponse
+import com.app.edonymyeon.data.service.client.calladapter.ApiResponse
 import retrofit2.http.GET
 import retrofit2.http.Query
