+++ android/app/src/main/java/com/app/edonymyeon/data/service/PreferenceService.kt
@@ -1,8 +1,8 @@
 package com.app.edonymyeon.data.service
 
-import com.app.edonymyeon.data.common.ApiResponse
 import com.app.edonymyeon.data.dto.request.NotificationPreferenceRequest
 import com.app.edonymyeon.data.dto.response.NotificationPreferenceResponse
+import com.app.edonymyeon.data.service.client.calladapter.ApiResponse
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
