+++ android/app/src/main/java/com/app/edonymyeon/data/service/ReportService.kt
@@ -1,7 +1,7 @@
 package com.app.edonymyeon.data.service
 
-import com.app.edonymyeon.data.common.ApiResponse
 import com.app.edonymyeon.data.dto.request.ReportRequest
+import com.app.edonymyeon.data.service.client.calladapter.ApiResponse
 import retrofit2.http.Body
 import retrofit2.http.POST
